requirements.txt
asyncio
pyrogram
python-dotenv
aiogram
sqlalchemy